refactor: streamline researcher dashboard and study proposal components
