+++ src/app/researcher/study/request/[orgSlug]/study-proposal.tsx
@@ -8,7 +8,7 @@ import { useForm } from '@mantine/form'
 import { studyProposalFormSchema, codeFilesSchema, StudyProposalFormValues } from './study-proposal-form-schema'
 import { StudyProposalForm } from './study-proposal-form'
 import { UploadStudyJobCode } from './upload-study-job-code'
-import { useMutation } from '@tanstack/react-query'
+import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { onCreateStudyAction, onDeleteStudyAction } from './actions'
 import { useRouter } from 'next/navigation'
 import { zodResolver } from '@/components/common'
@@ -105,6 +105,8 @@ export const StudyProposal: React.FC<{ orgSlug: string }> = ({ orgSlug }) => {
         ],
     })
 
+    const queryClient = useQueryClient()
+
     const { isPending, mutate: createStudy } = useMutation({
         mutationFn: async (formValues: StudyProposalFormValues) => {
             // Don't send any actual files to the server action, because they can't handle the file sizes
@@ -153,8 +155,8 @@ export const StudyProposal: React.FC<{ orgSlug: string }> = ({ orgSlug }) => {
                     'Your proposal has been successfully submitted to the reviewing organization. Check your dashboard for status updates.',
                 color: 'green',
             })
+            queryClient.invalidateQueries({ queryKey: ['researcher-studies'] })
             router.push(`/researcher/dashboard`)
-            window.location.assign(`/researcher/dashboard`)
         },
         onError: async (error, _, context: { studyId: string; studyJobId: string } | undefined) => {
             console.error(error)
