+++ src/app/reviewer/[orgSlug]/study/[studyId]/review/job-review-buttons.tsx
@@ -5,9 +5,9 @@ import { approveStudyJobFilesAction, rejectStudyJobFilesAction } from '@/server/
 import type { LatestJobForStudy } from '@/server/db/queries'
 import { Button, Group, Text, useMantineTheme } from '@mantine/core'
 import { CheckCircleIcon, XCircleIcon } from '@phosphor-icons/react/dist/ssr'
-import { useMutation } from '@tanstack/react-query'
+import { useMutation, useQueryClient } from '@tanstack/react-query'
 import dayjs from 'dayjs'
-import { useParams } from 'next/navigation'
+import { useParams, useRouter } from 'next/navigation'
 
 export const JobReviewButtons = ({
     job,
@@ -17,7 +17,9 @@ export const JobReviewButtons = ({
     decryptedResults?: JobFileInfo[]
 }) => {
     const theme = useMantineTheme()
+    const router = useRouter()
     const { orgSlug } = useParams<{ orgSlug: string }>()
+    const queryClient = useQueryClient()
 
     const {
         mutate: updateStudyJob,
@@ -44,7 +46,8 @@ export const JobReviewButtons = ({
         },
         onError: reportMutationError('Failed to update study job status'),
         onSuccess: () => {
-            window.location.assign(`/reviewer/${orgSlug}/dashboard`)
+            queryClient.invalidateQueries({ queryKey: ['org-studies', orgSlug] })
+            router.push(`/reviewer/${orgSlug}/dashboard`)
         },
     })
